pyproject.toml
[project]
name = "otp-verification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.2.1",
]
